Fix halt height types

+++ heimdall-migration-monitor/monitor.go
@@ -55,7 +55,7 @@ func heimdallHaltHeightMonitor(heimdallUrl string) {
 
 		var haltHeightResponse struct {
 			Height string `json:"height"`
-			Result int    `json:"result"`
+			Result int64  `json:"result"`
 		}
 
 		err = json.NewDecoder(resp.Body).Decode(&haltHeightResponse)
@@ -71,7 +71,7 @@ func heimdallHaltHeightMonitor(heimdallUrl string) {
 			continue
 		}
 
-		if haltHeightResponse.Result > int(currentHeight) && haltHeightResponse.Result-int(currentHeight) < 100 {
+		if haltHeightResponse.Result > currentHeight && haltHeightResponse.Result-currentHeight < 100 {
 			IsHFApproaching = true
 		} else {
 			IsHFApproaching = true
